marco-voice-engine/pyproject.toml
[project]
name = "marco-voice-engine"
version = "0.1.0"
description = "Skeleton for a modular voice brand engine"
readme = "README.md"
authors = [{name = "Codex"}]
requires-python = ">=3.11"
dependencies = [
    "httpx",
    "faiss-cpu",
    "python-dotenv==1.0.1",
    "python-telegram-bot==21.6",
]

[project.optional-dependencies]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/marco_voice_engine"]

[tool.hatch.build.targets.sdist]
include = ["src", "README.md", "pyproject.toml", "data"]
